Implemented a replace priority system for clues
Only affects GameController. Now there's a priorityQueque that sets the priority depending on a new data on the int saved in the userData of each textField.
Some refactoring in the conditionals and resolvabilityInformationLabel system.
The help button now checks if there's data in the queque, and picks from it the position to replace. If it is empty, it has the same behaviour previous to the update.
Now the program only checks Solvability when the input the user did is actually a valid number in the sudoku.
Fixed the color changing of the cells when a input is validated (Previous to this commit the borders were broken)